# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Expense Reimbursement System API (Synced Schemas)
  version: 2.0.0
  description: |
    API definitions generated from current Pydantic schemas and routers.
    Request and response bodies strictly match the Python schema definitions used by the FastAPI routers.
servers:
  - url: https://localhost:8081/api/v1
    description: Local test server

security:
  - firebaseAuth: []

components:
  securitySchemes:
    firebaseAuth:
      type: http
      scheme: bearer
      bearerFormat: FirebaseIDToken
      description: |
        Provide Firebase ID Token in the Authorization header.
        Example: `Bearer eyJhbGciOiJ...`

  schemas:
    EmployeeCreate:
      type: object
      required: [authentication_id, email, name, position, role]
      properties:
        authentication_id: { type: string }
        email: { type: string, format: email }
        name: { type: string }
        position: { type: string }
        role: { type: string, enum: ["employee", "admin"], default: "employee" }
        bank_account: { type: string, nullable: true }
        notes: { type: string, nullable: true }

    EmployeeUpdate:
      # Same shape as EmployeeCreate (used for PATCH)
      allOf:
        - $ref: '#/components/schemas/EmployeeCreate'

    EmployeeOut:
      allOf:
        - $ref: '#/components/schemas/EmployeeCreate'
        - type: object
          required: [employee_id]
          properties:
            employee_id: { type: string }

    ExpenseCreate:
      type: object
      required: [employee_id, amount, category]
      properties:
        employee_id: { type: string }
        amount: { type: number }
        category: { type: string }
        business_justification: { type: string, default: "" }

    ExpenseOut:
      type: object
      required: [expense_id, status, submitted_at]
      properties:
        expense_id: { type: string }
        status: { type: string, enum: ["pending", "approved", "rejected"], default: "pending" }
        submitted_at: { type: string, format: date-time }

    ExpenseUpdate:
      type: object
      required: [expense_id, date_of_expense, employee_id, amount, business_justification, category]
      properties:
        expense_id: { type: string }
        date_of_expense: { type: string, format: date-time }
        employee_id: { type: string }
        amount: { type: number }
        business_justification: { type: string }
        category: { type: string, enum: ["Travel", "Meals", "Conference", "Other"], default: "Other" }

        status: { type: string, enum: ["pending", "approved", "rejected"], default: "pending" }
        decision_actor: { type: string, enum: ["AI", "Human"], default: "AI" }
        decision_reason: { type: string, nullable: true }

        receipt_path: { type: string, nullable: true }

        submitted_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }

    AuditLogCreate:
      type: object
      required: [actor]
      properties:
        actor: { type: string }
        log: { type: string, nullable: true, default: "" }

    AuditLogOut:
      allOf:
        - $ref: '#/components/schemas/AuditLogCreate'
        - type: object
          required: [timestamp]
          properties:
            timestamp: { type: string, format: date-time }

    ReceiptUrl:
      type: object
      required: [expense_id, url]
      properties:
        expense_id: { type: string }
        url: { type: string, format: uri }

paths:
  /employees:
    get:
      tags: [Employee]
      summary: List all employees
      security:
        - firebaseAuth: []
      responses:
        '200':
          description: List of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EmployeeOut'
    post:
      tags: [Employee]
      summary: Create new employee
      security:
        - firebaseAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeCreate'
      responses:
        '201':
          description: Employee created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeOut'

  /employees/{employee_id}:
    get:
      tags: [Employee]
      summary: Get employee by ID
      security:
        - firebaseAuth: []
      parameters:
        - name: employee_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Employee details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeOut'
        '404':
          description: Employee not found
    patch:
      tags: [Employee]
      summary: Update employee info
      security:
        - firebaseAuth: []
      parameters:
        - name: employee_id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeUpdate'
      responses:
        '200':
          description: Employee updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeOut'
    delete:
      tags: [Employee]
      summary: Delete employee
      security:
        - firebaseAuth: []
      parameters:
        - name: employee_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204':
          description: Employee deleted (no content)

  /expenses:
    get:
      tags: [Expense]
      summary: List all expenses
      security:
        - firebaseAuth: []
      responses:
        '200':
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExpenseOut'

    post:
      tags: [Expense]
      summary: Create new expense (multipart form with JSON string + receipt file)
      security:
        - firebaseAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [expense_data, receipt]
              properties:
                # Router expects a form field 'expense_data' containing the JSON string for ExpenseCreate
                expense_data:
                  type: string
                  description: JSON string matching the ExpenseCreate schema
                receipt:
                  type: string
                  format: binary
      responses:
        '201':
          description: Expense created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseOut'

  /expenses/{expense_id}:
    get:
      tags: [Expense]
      summary: Get expense by ID
      security:
        - firebaseAuth: []
      parameters:
        - name: expense_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Expense details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseOut'
        '404':
          description: Expense not found
    patch:
      tags: [Expense]
      summary: Update expense
      security:
        - firebaseAuth: []
      parameters:
        - name: expense_id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpenseUpdate'
      responses:
        '200':
          description: Expense updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseOut'
    delete:
      tags: [Expense]
      summary: Delete expense
      security:
        - firebaseAuth: []
      parameters:
        - name: expense_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204':
          description: Expense deleted (no content)

  /audit_logs:
    get:
      tags: [AuditLog]
      summary: List all audit logs
      security:
        - firebaseAuth: []
      responses:
        '200':
          description: List of audit logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuditLogOut'
    post:
      tags: [AuditLog]
      summary: Record new audit log entry
      security:
        - firebaseAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuditLogCreate'
      responses:
        '201':
          description: Audit log created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLogOut'

  /documents/receipt/{expense_id}:
    get:
      tags: [Documents]
      summary: Get a temporary URL to view a receipt PDF for an expense
      security:
        - firebaseAuth: []
      parameters:
        - name: expense_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Temporary receipt URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiptUrl'
        '404':
          description: Expense or receipt not found